Push cursed token to an ephemeral branch
This is another attempt at making the token easier to access for users
(as current approach requires a GitHub token and using the GitHub API).

* Remove the previous attempt (GH Pages publishing): publishing this
  often does not work in practice
* Commit each token in git instead, use an ephemeral branch that gets
  overwritten on each push

Users *could* access the token by simply downloading
https://raw.githubusercontent.com/sigstore-conformance/extremely-dangerous-public-oidc-beacon/current-token/oidc-token.txt
Unfortunately the raw access is aggressively cached so this is not
practical: using git is the only way to get timely access to the latest
token.

Closes #8.
This may also fix #4 ("Figure out a better way to keep the beacon alive").

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>