Publish OIDC beacon in branch (#18)
* Push cursed token to an ephemeral branch

This is another attempt at making the token easier to access for users
(as current approach requires a GitHub token and using the GitHub API).

* Remove the previous attempt (GH Pages publishing): publishing this
  often does not work in practice
* Commit each token in git instead, use an ephemeral branch that gets
  overwritten on each push

Users *could* access the token by simply downloading
https://raw.githubusercontent.com/sigstore-conformance/extremely-dangerous-public-oidc-beacon/current-token/oidc-token.txt
Unfortunately the raw access is aggressively cached so this is not
practical: using git is the only way to get timely access to the latest
token.

Closes #8.
This may also fix #4 ("Figure out a better way to keep the beacon alive").

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

* Add an action to download the token

The purpose of this action is to handle retries: sometimes the token is
expired because GH Actions scheduling is not reliable enough. In that
case we want to wait a moment and try again.

Note that we're assuming pyjwt is already installed in system python.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

* README: Add info about the action

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

* Cleanup

* Fix the URL to actual repository
* Explicitly install pyjwt just in case
* Improve some comments, docs

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

* Remove pyjwt dependency

We can decode and parse the JWT ourselves like cavemen: We're not interested in
actually verifying the JWT, just sanity checking the expiry. This avoids
the pyjwt dependency.

Also add check=True to subprocess.run() for easier debugging.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

---------

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>